nitpicking

+++ arm.go
@@ -12,7 +12,7 @@ func getGOARM(goBinary string) uint8 {
 	for _, setting := range build.Settings {
 		if setting.Key == "GOARM" {
 			// the value is coming from the linker, so it should be safe to convert
-			return uint8(setting.Value[0] - '0')
+			return setting.Value[0] - '0'
 		}
 	}
 	return 0
